chore: standarize how predictions are made
