+++ gragod/predictions/spot.py
